obj transformation logic works but not based on manipulator selection

+++ CS450AssignmentThree/include/mat.h
@@ -578,6 +578,7 @@ mat4 transpose( const mat4& A ) {
 inline
 vec4 mvmult( const mat4& a, const vec4& b )
 {
+
     vec4 c;
     int i, j;
     for(i=0; i<4; i++) {
@@ -637,9 +638,9 @@ inline
 mat4 Translate( const GLfloat x, const GLfloat y, const GLfloat z )
 {
     mat4 c;
-    c[0][3] = x;
-    c[1][3] = y;
-    c[2][3] = z;
+    c[0][3] = GLfloat(x);
+    c[1][3] = GLfloat(y);
+    c[2][3] = GLfloat(z);
     return c;
 }
